// EmployeeManager.ClientApp/src/app/core/api/generated/model/salaryDto.ts
/**
 * EmployeeManager.WebAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface SalaryDto { 
    id?: number;
    amount?: number;
    from?: string;
    to?: string | null;
    employeeId?: number;
}